Поменял в ценах int на BigDecimal с занятия

+++ api/src/main/java/ru/yajaneya/Spring2Geekbrains/api/carts/CartDto.java
@@ -1,15 +1,16 @@
 package ru.yajaneya.Spring2Geekbrains.api.carts;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class CartDto {
     private List<CartItemDto> items;
-    private int totalPrice;
+    private BigDecimal totalPrice;
 
     public CartDto() {
     }
 
-    public CartDto(List<CartItemDto> items, int totalPrice) {
+    public CartDto(List<CartItemDto> items, BigDecimal totalPrice) {
         this.items = items;
         this.totalPrice = totalPrice;
     }
@@ -22,11 +23,11 @@ public class CartDto {
         this.items = items;
     }
 
-    public int getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(int totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
 }
