+++ api/src/main/java/ru/yajaneya/Spring2Geekbrains/api/core/OrderDto.java
@@ -1,19 +1,20 @@
 package ru.yajaneya.Spring2Geekbrains.api.core;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class OrderDto {
     private Long id;
     private String username;
     private List<OrderItemDto> items;
-    private Integer totalPrice;
+    private BigDecimal totalPrice;
     private String address;
     private String phone;
 
     public OrderDto() {
     }
 
-    public OrderDto(Long id, String username, List<OrderItemDto> items, Integer totalPrice, String address, String phone) {
+    public OrderDto(Long id, String username, List<OrderItemDto> items, BigDecimal totalPrice, String address, String phone) {
         this.id = id;
         this.username = username;
         this.items = items;
@@ -46,11 +47,11 @@ public class OrderDto {
         this.items = items;
     }
 
-    public Integer getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(Integer totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
