+++ cart-service/src/test/java/ru/yajaneya/Spring2Geekbrains/cart/CartTest.java
@@ -12,6 +12,7 @@ import ru.yajaneya.Spring2Geekbrains.api.core.ProductDto;
 import ru.yajaneya.Spring2Geekbrains.cart.integretions.ProductsServiceIntegration;
 import ru.yajaneya.Spring2Geekbrains.cart.services.CartService;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 @SpringBootTest
@@ -28,12 +29,12 @@ public class CartTest {
         ProductDto product1 = new ProductDto();
         product1.setId(1L);
         product1.setTitle("Milk");
-        product1.setPrice(50);
+        product1.setPrice(BigDecimal.valueOf(50.00));
         product1.setCategory("Milk");
         ProductDto product2 = new ProductDto();
         product2.setId(2L);
         product2.setTitle("Cream");
-        product2.setPrice(150);
+        product2.setPrice(BigDecimal.valueOf(150.00));
         product2.setCategory("Milk");
         Mockito.doReturn(Optional.of(product1)).when(productsService).findById(1L);
         Mockito.doReturn(Optional.of(product2)).when(productsService).findById(2L);
